packages/mcp-server/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mikrotik-mcp-server"
version = "0.1.0"
description = "MCP server for MikroTik RouterOS"
requires-python = ">=3.11"
dependencies = [
  "jq>=1,<2",
  "mcp>=1,<2",
  "python-dotenv>=1,<2",
]

[project.optional-dependencies]
test = [
  "pytest>=8,<9",
  "pytest-asyncio>=0.23,<1",
  "pytest-socket>=0.7,<1",
]

[tool.setuptools]
package-dir = {"" = "src"}
py-modules = ["main", "server", "client"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = "--disable-socket"
testpaths = ["tests"]
